Fix NotEq{"x": []interface{}} generation
This was resolving to FALSE, but should always be TRUE.

The only portable way I found to do this was to replace the expression
with a portable TRUE (1=1). Also changed Eq{"x": []interface{}} to a
portable FALSE (1=0) for consistency to avoid nasty surprises around
returning NULL instead of a boolean.

Fixes #99